--- Bodu.Security.Cryptography/test/Security.Cryptography/AsconHashA256Tests.Dispose.cs
@@ -1,49 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="AsconHashA256Tests.Dispose.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-namespace Bodu.Security.Cryptography;
-
-public partial class AsconHashA256Tests
-{
-    /// <summary>
-    /// Verifies that calling <see cref="AsconHashA256.Dispose" /> twice on the same instance is
-    /// idempotent and does not throw.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenCalledTwice_ShouldNotThrow()
-    {
-        var algorithm = new AsconHashA256();
-        algorithm.Dispose();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Second Dispose on AsconHashA256 threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-
-    /// <summary>
-    /// Verifies that disposing a freshly-constructed <see cref="AsconHashA256" /> instance — one
-    /// that has never had any property accessed or hashing performed — completes without throwing.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenInstanceUntouched_ShouldNotThrow()
-    {
-        var algorithm = new AsconHashA256();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Disposing an untouched AsconHashA256 instance threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-}
