--- Bodu.Security.Cryptography/test/Security.Cryptography/Blake2sTests.Dispose.cs
@@ -1,50 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="Blake2sTests.Dispose.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-namespace Bodu.Security.Cryptography;
-
-public partial class Blake2sTests
-{
-    /// <summary>
-    /// Verifies that calling <see cref="Blake2s.Dispose" /> twice on the same instance is
-    /// idempotent and does not throw.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenCalledTwice_ShouldNotThrow()
-    {
-        var algorithm = new Blake2s();
-        algorithm.Dispose();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Second Dispose on Blake2s threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-
-    /// <summary>
-    /// Verifies that disposing a freshly-constructed <see cref="Blake2s" /> instance — one that
-    /// has never had its key, hash size, or any other property accessed — completes without
-    /// throwing.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenInstanceUntouched_ShouldNotThrow()
-    {
-        var algorithm = new Blake2s();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Disposing an untouched Blake2s instance threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-}
