--- Bodu.Security.Cryptography/test/Security.Cryptography/CubeHashTests.Dispose.cs
@@ -1,49 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="CubeHashTests.Dispose.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-namespace Bodu.Security.Cryptography;
-
-public partial class CubeHashTests
-{
-    /// <summary>
-    /// Verifies that calling <see cref="CubeHash.Dispose" /> twice on the same instance is
-    /// idempotent and does not throw.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenCalledTwice_ShouldNotThrow()
-    {
-        var algorithm = new CubeHash();
-        algorithm.Dispose();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Second Dispose on CubeHash threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-
-    /// <summary>
-    /// Verifies that disposing a freshly-constructed <see cref="CubeHash" /> instance — one that
-    /// has never had any property accessed or hashing performed — completes without throwing.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenInstanceUntouched_ShouldNotThrow()
-    {
-        var algorithm = new CubeHash();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Disposing an untouched CubeHash instance threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-}
