+++ Bodu.Security.Cryptography/test/Security.Cryptography/HashAlgorithmTests.ComputeHash.cs
@@ -527,5 +527,27 @@ public abstract partial class HashAlgorithmTests<TTest, TAlgorithm, TVariant>
         });
     }
 
-  
+    /// <summary>
+    /// Verifies that two consecutive <see cref="HashAlgorithm.ComputeHash(byte[])" /> calls
+    /// against the same instance produce identical digests for identical input — guarding against
+    /// state leakage across reuse. Skipped for one-shot algorithms that report
+    /// <see cref="HashAlgorithm.CanReuseTransform" /> as <see langword="false" />.
+    /// </summary>
+    [TestMethod]
+    public void ComputeHash_WhenInvokedRepeatedly_ShouldYieldIdenticalDigest()
+    {
+        using var algorithm = CreateAlgorithm();
+        if (!algorithm.CanReuseTransform)
+        {
+            Assert.Inconclusive($"{typeof(TAlgorithm).Name} reports CanReuseTransform=false; same-instance reuse is not supported.");
+            return;
+        }
+
+        byte[] input = Enumerable.Range(0, 256).Select(i => (byte)i).ToArray();
+
+        byte[] first = algorithm.ComputeHash(input);
+        byte[] second = algorithm.ComputeHash(input);
+
+        CollectionAssert.AreEqual(first, second);
+    }
 }
