--- Bodu.Security.Cryptography/test/Security.Cryptography/SipHashTests.Dispose.cs
@@ -1,30 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="SipHashTests.Dispose.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-namespace Bodu.Security.Cryptography;
-
-public abstract partial class SipHashTests<TTest, TAlgorithm>
-{
-    /// <summary>
-    /// Verifies that calling <see cref="System.Security.Cryptography.HashAlgorithm.Dispose()" /> twice on a SipHash instance is
-    /// idempotent and does not throw.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenCalledTwice_ShouldNotThrow()
-    {
-        var algorithm = CreateAlgorithm();
-        algorithm.Dispose();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Second Dispose on {typeof(TAlgorithm).Name} threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-}
