+++ Bodu.Security.Cryptography/test/Security.Cryptography/TigerTests.AlgorithmName.cs
@@ -1,10 +1,8 @@
-﻿// ---------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------
 // <copyright file="TigerTests.AlgorithmName.cs" company="PlaceholderCompany">
 //     Copyright (c) PlaceholderCompany. All rights reserved.
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
-
-using System.Security.Cryptography;
 
 namespace Bodu.Security.Cryptography;
 
@@ -20,21 +18,5 @@ public partial class TigerTests
         using var algorithm = CreateAlgorithm(variant);
 
         Assert.AreEqual($"Tiger/{algorithm.HashSize}", algorithm.AlgorithmName);
-    }
-
-    /// <summary>
-    /// Verifies that reading <see cref="Tiger.AlgorithmName" /> after <see cref="HashAlgorithm.Dispose()" />
-    /// throws <see cref="ObjectDisposedException" /> and not a generic <see cref="NullReferenceException" />.
-    /// </summary>
-    [TestMethod]
-    public void AlgorithmName_WhenAccessedAfterDispose_ShouldThrowExactly()
-    {
-        var algorithm = CreateAlgorithm();
-        algorithm.Dispose();
-
-        Assert.ThrowsExactly<ObjectDisposedException>(() =>
-        {
-            _ = algorithm.AlgorithmName;
-        });
     }
 }
