--- Bodu.Security.Cryptography/test/Security.Cryptography/TigerTests.Dispose.cs
@@ -1,29 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="TigerTests.Dispose.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-namespace Bodu.Security.Cryptography;
-
-public partial class TigerTests
-{
-    /// <summary>
-    /// Verifies that calling <see cref="Tiger.Dispose" /> a second time is idempotent and does not throw.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenCalledTwice_ShouldNotThrow()
-    {
-        var algorithm = CreateAlgorithm();
-        algorithm.Dispose();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Second Dispose on Tiger threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-}
