--- Bodu.Security.Cryptography/test/Security.Cryptography/TigerTests.Initialize.cs
@@ -1,29 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="TigerTests.Initialize.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-using System.Security.Cryptography;
-
-namespace Bodu.Security.Cryptography;
-
-public partial class TigerTests
-{
-    /// <summary>
-    /// Verifies that calling <see cref="HashAlgorithm.Initialize" /> on a disposed instance
-    /// throws <see cref="ObjectDisposedException" /> and not <see cref="NullReferenceException" />
-    /// from the residual buffer access in the base class.
-    /// </summary>
-    [TestMethod]
-    public void Initialize_WhenCalledAfterDispose_ShouldThrowExactly()
-    {
-        var algorithm = CreateAlgorithm();
-        algorithm.Dispose();
-
-        Assert.ThrowsExactly<ObjectDisposedException>(() =>
-        {
-            algorithm.Initialize();
-        });
-    }
-}
