--- Bodu.Security.Cryptography/test/Security.Cryptography/WhirlpoolTests.ComputeHash.cs
@@ -1,30 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="WhirlpoolTests.ComputeHash.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-using System.Security.Cryptography;
-
-namespace Bodu.Security.Cryptography;
-
-public partial class WhirlpoolTests
-{
-    /// <summary>
-    /// Verifies that two consecutive <see cref="HashAlgorithm.ComputeHash(byte[])" /> calls
-    /// against the same <see cref="Whirlpool" /> instance produce identical digests for identical
-    /// input — guarding against state leakage across reuse and confirming that
-    /// <see cref="Whirlpool.Version" /> selection persists across hashes.
-    /// </summary>
-    [TestMethod]
-    public void ComputeHash_WhenInvokedRepeatedly_ShouldYieldIdenticalDigest()
-    {
-        using var algorithm = new Whirlpool();
-        byte[] input = Enumerable.Range(0, 256).Select(i => (byte)i).ToArray();
-
-        byte[] first = algorithm.ComputeHash(input);
-        byte[] second = algorithm.ComputeHash(input);
-
-        CollectionAssert.AreEqual(first, second);
-    }
-}
