--- Bodu.Security.Cryptography/test/Security.Cryptography/WhirlpoolTests.Dispose.cs
@@ -1,49 +1,0 @@
-// ---------------------------------------------------------------------------------------------------------------
-// <copyright file="WhirlpoolTests.Dispose.cs" company="PlaceholderCompany">
-//     Copyright (c) PlaceholderCompany. All rights reserved.
-// </copyright>
-// ---------------------------------------------------------------------------------------------------------------
-
-namespace Bodu.Security.Cryptography;
-
-public partial class WhirlpoolTests
-{
-    /// <summary>
-    /// Verifies that calling <see cref="Whirlpool.Dispose" /> twice on the same instance is
-    /// idempotent and does not throw.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenCalledTwice_ShouldNotThrow()
-    {
-        var algorithm = new Whirlpool();
-        algorithm.Dispose();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Second Dispose on Whirlpool threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-
-    /// <summary>
-    /// Verifies that disposing a freshly-constructed <see cref="Whirlpool" /> instance — one that
-    /// has never had any property accessed or hashing performed — completes without throwing.
-    /// </summary>
-    [TestMethod]
-    public void Dispose_WhenInstanceUntouched_ShouldNotThrow()
-    {
-        var algorithm = new Whirlpool();
-
-        try
-        {
-            algorithm.Dispose();
-        }
-        catch (Exception ex)
-        {
-            Assert.Fail($"Disposing an untouched Whirlpool instance threw {ex.GetType().Name}: {ex.Message}");
-        }
-    }
-}
